% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_aft_fit.R, R/methods_print.R
\name{iv_aft_fit}
\alias{iv_aft_fit}
\alias{coef.cTSLS}
\alias{print.cTSLS}
\alias{summary.cTSLS}
\alias{vcov.cTSLS}
\alias{predict.cTSLS}
\title{Fit a two‐stage IV accelerated‐failure‐time model}
\usage{
iv_aft_fit(formula_stage2, formula_stage1, data, maxit = 10, Naive = FALSE)

\method{coef}{cTSLS}(object, stage = c("second", "first", "both"), ...)

\method{print}{cTSLS}(x, ...)

\method{summary}{cTSLS}(object, ...)

\method{vcov}{cTSLS}(object, stage = c("second", "first", "full"), naive = FALSE, ...)

\method{predict}{cTSLS}(object, newdata = NULL, type = c("response", "link"), ...)
}
\arguments{
\item{formula_stage2}{A second‐stage formula, e.g. \code{Surv(time, status) ~ W + X1 + X2}}

\item{formula_stage1}{A first‐stage formula, e.g. \code{W ~ Z1 + Z2}}

\item{data}{A data.frame containing all variables in the formulas}

\item{maxit}{Integer; maximum number of iterations.}

\item{Naive}{Whether use the naive sandwich estimator of variance}

\item{stage}{Which covariance to return: "second" (default), "first", or "full".}

\item{naive}{Logical; if TRUE, return the naive sandwich VCOV (if stored).}
}
\value{
An object of class \code{cTSLS} with components:
  \describe{
    \item{theta}{Second‐stage coefficients (including intercept).}
    \item{var}{Placeholder for sandwich variance matrix (currently \code{NULL}).}
    \item{first_stage}{Output list from \code{first_stage()}.}
    \item{second_stage}{Output list from \code{second_stage_wls()}.}
    \item{call}{The matched call.}
  }
}
\description{
Fit a two‐stage IV accelerated‐failure‐time model
}
