% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_stage.R
\name{first_stage}
\alias{first_stage}
\title{First‐stage OLS for the endogenous exposure (via lm)}
\usage{
first_stage(W, Cov, add_intercept = TRUE)
}
\arguments{
\item{W}{Numeric vector of the endogenous variable.}

\item{Cov}{Numeric matrix or data.frame of instrument(s).}

\item{add_intercept}{Logical; include intercept? Default TRUE.}
}
\value{
A list with \code{beta}, \code{fitted}, \code{resid}, and the lm \code{model}.
}
\description{
First‐stage OLS for the endogenous exposure (via lm)
}
