% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.R
\name{synthetic_leurgans}
\alias{synthetic_leurgans}
\title{Leurgans synthetic outcome for right‐censoring}
\usage{
synthetic_leurgans(time, status)
}
\arguments{
\item{time}{Numeric vector of observed times.}

\item{status}{Integer censoring indicator (1=event, 0=censored).}
}
\value{
A list with \code{Ystar}, \code{sigma2}, \code{Ghat}.
}
\description{
Leurgans synthetic outcome for right‐censoring
}
